DmdAnal: When aborting fixed-point-iteration, do not forget strict variables
When fixed-point iteration does not terminate, we conservatively delete
their strictness signatures (set them to nopSig). But this loses the
information on how its strict free variables are used!

Lazily used variables already escape via lazy_fvs. This patch ensures
that in the case of an aborted fixed-point iteration, also the strict
variables are put there (with a conservative demand of topDmd).